==> src/2_K vecinos SVM proyecto.ipynb <==
import pandas as pd
import numpy as np
import sklearn.metrics as mtrcs
import math
import random  #para el randomize

#Scikit
from scipy.stats import norm
from scipy import stats as st
from sklearn.metrics import confusion_matrix
from sklearn.cross_validation import train_test_split
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn.naive_bayes import GaussianNB
from sklearn import svm
from sklearn import tree
import os #para exportar el pdf del tree
from sklearn.datasets import load_iris   #para hacer load de datas iris

#Éste es el objetivo de este experimento:
from sklearn.neighbors import NearestNeighbors  #for unsupervised
from sklearn.neighbors import KNeighborsRegressor #for supervised (the case now :) )

credit = pd.read_csv("credit4.csv")
print(credit.head())

# K vecinos

#datosnormales
X_train, X_test,Y_train,Y_test = train_test_split(credit.iloc[:,0:11],credit[['default.payment.next.month']], train_size=0.75)
Xf=np.asarray(X_train)
Yf=np.asarray(Y_train)
Xftes = np.asarray(X_test)
Yftes = np.asarray(Y_test)

neighDivide = KNeighborsRegressor(n_neighbors=1)
neighDivide.fit(X_train, Y_train)
predictNeigh_test = neighDivide.predict(X_test)

print 'confusionmatrix = ', confusion_matrix(Y_test, predictNeigh_test)
print 'precision = ', mtrcs.precision_score(Y_test, predictNeigh_test)
print 'Accuracy Score = ', mtrcs.accuracy_score(Y_test, predictNeigh_test)

# SVM

#penalty = 0.3  #esta variable determina qué tanto "margen dejar"
clf = svm.SVC(kernel="rbf",probability=True)
h = clf.fit(Xf[:,1:11], Yf).predict(Xf[:,1:11])
g = clf.predict(Xftes[:,1:11])
# "probabilities"
#probas_X_train = clf.predict_proba(Xf[:,1:11])
#probas_X_test = clf.predict_proba(Xftes[:,1:11])
#y_predSVM = clf.predict(Xftes[:,1:11])
print 'confusionmatrix = ', confusion_matrix(Yftes, g)
print 'precision = ', mtrcs.precision_score(Yftes, g)
print 'Accuracy Score = ', mtrcs.accuracy_score(Yftes, g)

#guardando el vector de probas con indices 
print(Xf[:,0],probas_X_train2)
np.savetxt("cord_Train.csv",Xf[:,0],delimiter=",")
np.savetxt("cord_test.csv",Xftes[:,0],delimiter=",")
np.savetxt("prob_train.csv",probas_X_train,delimiter=",")
np.savetxt("prob_test.csv",probas_X_test,delimiter=",")
