==> src/3_Red Neuronal.ipynb <==
import tensorflow as tf
import numpy as np
from sklearn.cross_validation import train_test_split
#import pandas as pd
import matplotlib.pyplot as plt
import pandas as pd
%matplotlib inline

credit2 = pd.read_csv("credit4.csv")
credit2.head()

X_train, X_test,Y_train,Y_test = train_test_split(credit2.iloc[:,1:12],credit2[['default.payment.next.month']], train_size=0.75)
Xf=np.asarray(X_train)
Yf=np.asarray(Y_train)
Xftes = np.asarray(X_test)
Yftes = np.asarray(Y_test)
X_train.head()

#declaracion de palabras
input_size=11
output_layer_size=1
#Los placeholders son para los datos de entrada
x = tf.placeholder(tf.float32, [None, input_size])
y_ = tf.placeholder(tf.float32, [None, output_layer_size]) #la y real

#La variables para lo que se va a ir calculando y modificando en el camino
#W_layer1=tf.Variable(tf.random_uniform([input_size,output_layer_size], -1, 1), name="W_layer1")
#b_layer1 = tf.Variable(tf.zeros([output_layer_size]), name="b_layer1")

W_layer1 = tf.Variable(tf.random_uniform([11, 8], -1, 1), 
                       name = "W_layer1")
b_layer1 = tf.Variable(tf.zeros([8]), 
                       name = "b_layer1")

W_layer2 = tf.Variable(tf.random_uniform([8, 6], -1, 1), 
                       name = "W_layer2")
b_layer2 = tf.Variable(tf.zeros([6]), 
                       name = "b_layer2")

W_layer3 = tf.Variable(tf.random_uniform([6, 1], -1, 1), 
                       name = "W_layer3")
b_layer3 = tf.Variable(tf.zeros([1]), 
                       name = "b_layer3")
#operaciones de grafo
hd1 = tf.nn.sigmoid( tf.matmul(x, W_layer1) + b_layer1)
hd2 = tf.nn.sigmoid( tf.matmul(hd1, W_layer2) + b_layer2)
y = tf.nn.sigmoid( tf.matmul(hd2, W_layer3) + b_layer3) 
#y = tf.nn.sigmoid(tf.matmul(x,W_layer1)+b_layer1) #la de nuestra
#lossfn = tf.reduce_mean(tf.reduce_sum((y_-y)**2)) #cuadratico
lossfn= tf.reduce_mean(( (y_ * tf.log(y)) +  
  ((1 - y_) * tf.log(1.0 - y)) ) * -1) # cross entropy
train_step = tf.train.GradientDescentOptimizer(0.05).minimize(lossfn) #ojo si disminuye baja el accuracy 

init = tf.initialize_all_variables()
sess = tf.Session() # tipo de sesion, puede ser interactiva

#correr programa
sess.run(init)
for i in range(30000):
  sess.run(train_step, feed_dict={x: Xf, y_: Yf})

#imprimir alguna variable
print(sess.run(y, feed_dict={x: Xf, y_: Yf}))
print(sess.run(W_layer1, feed_dict={x: Xf, y_: Yf}))

#cuantificar error
correct_prediction = tf.equal(tf.round(y),y_) # aca estams definiendo otras operacioes
accuracy = tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
print(sess.run(accuracy, feed_dict={x: Xf, y_: Yf})) # las corremos
